improved deb build script

+++ build-deb.sh
@@ -7,16 +7,16 @@ fi
 
 set -ex
 
-python setup.py clean
-python setup.py sdist
-python setup.py clean
-
-VERSION=$(python -c "from ${1} import __version__; print __version__")
-
 typeset -l lowername
 lowername=$1
-mv dist/${1}-${VERSION}.tar.gz ../${lowername}_${VERSION}.orig.tar.gz
+VERSION=$(python -c "from ${1} import __version__; print __version__")
 
-rm -rf dist build
+if [ ! -f ../${lowername}_${VERSION}.orig.tar.gz ]; then
+    python setup.py clean
+    python setup.py sdist
+    python setup.py clean
+    mv dist/${1}-${VERSION}.tar.gz ../${lowername}_${VERSION}.orig.tar.gz
+    rm -rf dist build
+fi
 
 fakeroot dpkg-buildpackage -i.git -S -sa
